Update encoding tests
