+++ test/functional/encoding/mocks/with-charset.html
@@ -9,5 +9,6 @@
     <div id="special-characters-ukrainian">Слава Україні!</div>
     <div id="special-characters-chinese">加入网站</div>
     <div id="special-characters-ukrainian">Обладнання та ПЗ</div>
+    <div id="special-characters-french">PAR PASSION DU VÉLO</div>
 </body>
 </html>
