+++ test/functional/encoding/mocks/without-charset.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Test</title>
+</head>
+<body>
+    <div id="special-characters-korean">저는 7년 동안 한국에서 살았어요.</div>
+    <div id="special-characters-ukrainian">Слава Україні!</div>
+    <div id="special-characters-chinese">加入网站</div>
+    <div id="special-characters-ukrainian">Обладнання та ПЗ</div>
+    <div id="special-characters-french">PAR PASSION DU VÉLO</div>
+</body>
+</html>
